+++ util.go
@@ -11,7 +11,7 @@ import (
 )
 
 func BinaryBytes(o interface{}) []byte {
-	w, n, err := new(bytes.Buffer), new(int64), new(error)
+	w, n, err := new(bytes.Buffer), new(int), new(error)
 	WriteBinary(o, w, n, err)
 	if *err != nil {
 		PanicSanity(*err)
@@ -20,7 +20,7 @@ func BinaryBytes(o interface{}) []byte {
 }
 
 func JSONBytes(o interface{}) []byte {
-	w, n, err := new(bytes.Buffer), new(int64), new(error)
+	w, n, err := new(bytes.Buffer), new(int), new(error)
 	WriteJSON(o, w, n, err)
 	if *err != nil {
 		PanicSanity(*err)
@@ -59,7 +59,7 @@ func BinaryCompare(a, b interface{}) int {
 
 // NOTE: only use this if you need 32 bytes.
 func BinarySha256(o interface{}) []byte {
-	hasher, n, err := sha256.New(), new(int64), new(error)
+	hasher, n, err := sha256.New(), new(int), new(error)
 	WriteBinary(o, hasher, n, err)
 	if *err != nil {
 		PanicSanity(*err)
@@ -69,7 +69,7 @@ func BinarySha256(o interface{}) []byte {
 
 // NOTE: The default hash function is Ripemd160.
 func BinaryRipemd160(o interface{}) []byte {
-	hasher, n, err := ripemd160.New(), new(int64), new(error)
+	hasher, n, err := ripemd160.New(), new(int), new(error)
 	WriteBinary(o, hasher, n, err)
 	if *err != nil {
 		PanicSanity(*err)
